pyproject.toml
[project]
name = "emotion_detector"
version = "0.1.0"
description = "CNN-based facial emotion recognition using the FER2013 dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
